requirements.txt
django
libsass 
django-compressor 
django-sass-processor
djangorestframework
coverage
djangorestframework-xml